Kubernetes/CESMIIMQTTGraphQLDev/Client/requirements.txt
paho-mqtt
argparse
requests

